clever-backend/requirements.txt
requests
urllib3
python-firebase